+++ src/AboutPage.jsx
@@ -3,16 +3,7 @@ import React from 'react';
 import styled from '@emotion/styled';
 
 import PageWrapper from './PageWrapper';
-
-const ContentWrapper = styled.div({
-  borderStyle: 'solid none',
-  borderWidth: '1px',
-  borderColor: '#364f6b',
-  padding: '20px 20px',
-  margin: '30px auto',
-  backgroundColor: '#f5f5f5',
-  fontSize: '1.2em',
-});
+import ContentWrapper from './ContentWrapper';
 
 export default function AboutPage() {
   return (
