+++ src/ContentWrapper.jsx
@@ -1,0 +1,16 @@
+import styled from '@emotion/styled';
+
+const ContentWrapper = styled.div({
+  borderStyle: 'solid none',
+  borderWidth: '1px',
+  borderColor: '#364f6b',
+  padding: '20px 20px',
+  margin: '30px auto',
+  backgroundColor: '#f5f5f5',
+  fontSize: '1.2em',
+  '& h3': {
+    color: '#364f6b',
+  },
+});
+
+export default ContentWrapper;
