+++ src/Form.jsx
@@ -1,0 +1,8 @@
+import styled from '@emotion/styled';
+
+const Form = styled.div({
+  width: '90%',
+  margin: '0 auto',
+});
+
+export default Form;
