+++ src/LoginForm.jsx
@@ -1,32 +1,8 @@
 import React from 'react';
 
-import styled from '@emotion/styled';
-
+import Form from './Form';
+import FormGroup from './FormGroup';
 import Button from './Button';
-
-const Form = styled.div({
-  width: '90%',
-  margin: '0 auto',
-});
-
-const FormGroup = styled.div({
-  label: {
-    fontWeight: 'bold',
-    fontSize: '1.2em',
-    color: '#364f6b',
-  },
-  input: {
-    width: '100%',
-    padding: '12px 20px',
-    margin: '8px 0',
-    boxSizing: 'border-box',
-    border: '3px solid #364f6b',
-    '&:focus': {
-      border: '3px solid #3fc1c9',
-      outline: 'none',
-    },
-  },
-});
 
 const LoginForm = React.memo(({ fields, onChange, onSubmit }) => {
   const { email, password } = fields;
