+++ src/RestaurantDetail.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 
 import MenuItems from './MenuItems';
+import ContentWrapper from './ContentWrapper';
 
 export default function RestaurantDetail({ restaurant }) {
   const { name, address, menuItems } = restaurant;
@@ -8,13 +9,15 @@ export default function RestaurantDetail({ restaurant }) {
   return (
     <div>
       <h2>{name}</h2>
-      <p>
-        주소:
-        {' '}
-        {address}
-      </p>
-      <h3>메뉴</h3>
-      <MenuItems menuItems={menuItems} />
+      <ContentWrapper>
+        <p>
+          주소:
+          {' '}
+          {address}
+        </p>
+        <h3>메뉴</h3>
+        <MenuItems menuItems={menuItems} />
+      </ContentWrapper>
     </div>
   );
 }
